rust updates

+++ src/model/session/sequence.rs
@@ -3,19 +3,19 @@ pub struct Sequence<'a> {
 }
 
 impl Sequence<'a> {
-    pub fn Next() -> u64 {
+    pub fn Next() -> i32 {
 
     }
 }
 
 struct Node<'a> {
-    val: 'a u64,
+    val: 'a i32,
     left: Option<Box<Node<'a>>>,
     right: Option<Box<Node<'a>>>,
 }
 
 impl<'a> Node<'a> {
-    fn new(value: 'a u64) -> Node {
+    fn new(value: 'a i32) -> Node {
         Node {
             value: value,
             left: None,
@@ -23,7 +23,7 @@ impl<'a> Node<'a> {
         }
     }
 
-    pub fn insert(&mut self, new_val: 'a u64) {
+    pub fn insert(&mut self, new_val: 'a i32) {
         if self.val == new_val {
             return
         }
@@ -40,7 +40,7 @@ impl<'a> Node<'a> {
         }
     }
 
-    pub fn search(&self, target: u64) -> Option<i32> {
+    pub fn search(&self, target: i32) -> Option<i32> {
         match self.value {
             value if target == value => Some(value),
             value if target < value => self.left.as_ref()?.search(target),
